% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{check_for_empty_text}
\alias{check_for_empty_text}
\title{Check for Empty Text}
\usage{
check_for_empty_text(df)
}
\arguments{
\item{df}{A data frame of cleaned insolvency results.}
}
\description{
Checks if any scraped insolvency entries have empty text fields and stops execution if found.
}
