% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{clean_results}
\alias{clean_results}
\title{Clean Results}
\usage{
clean_results(df)
}
\arguments{
\item{df}{A data frame of scraped insolvency results.}
}
\value{
A cleaned and standardized data frame.
}
\description{
Cleans and standardizes the scraped insolvency data, including text manipulation and duplicate removal.
}
