% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{create_dates_to_crawl}
\alias{create_dates_to_crawl}
\title{Create Dates to Crawl}
\usage{
create_dates_to_crawl(df)
}
\arguments{
\item{df}{A data frame containing a 'datum' column (last crawled date).}
}
\value{
A vector of dates to be crawled.
}
\description{
Generates a sequence of dates to crawl, starting from the last crawled date up to today.
}
