% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{get_month_list}
\alias{get_month_list}
\title{Get Month List}
\usage{
get_month_list(dates)
}
\arguments{
\item{dates}{A vector of date objects or date strings.}
}
\value{
A vector of unique month strings.
}
\description{
Extracts unique month strings (YYYY-MM) from a vector of dates.
}
