% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{get_session_id}
\alias{get_session_id}
\title{Get Session ID}
\usage{
get_session_id(res)
}
\arguments{
\item{res}{An HTTP response object (from \code{httr::GET} or \code{httr::POST}).}
}
\value{
The 'jakarta.faces.ViewState' value.
}
\description{
Extracts the 'jakarta.faces.ViewState' (session ID) from an HTML response.
}
