% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{initiate_session}
\alias{initiate_session}
\title{Initiate Session}
\usage{
initiate_session(url)
}
\arguments{
\item{url}{The URL of the search page.}
}
\value{
The session ID obtained from the GET request.
}
\description{
Initiates a new web session by making a GET request to the search URL and extracting the session ID.
}
