% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{load_existing_data}
\alias{load_existing_data}
\title{Load Existing Insolvency Data}
\usage{
load_existing_data(data_path)
}
\arguments{
\item{data_path}{The file path to the Parquet dataset directory.}
}
\value{
A data frame containing the existing insolvency data.
}
\description{
Loads existing insolvency data from a Parquet dataset.
}
