% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inso_dataset.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load Insolvency Data from Raw Files}
\usage{
load_raw_data(data_path = "data-raw/inso")
}
\arguments{
\item{data_path}{The path to the directory containing the raw data files.}
}
\value{
A data frame containing the loaded insolvency data.
}
\description{
This function loads the insolvency raw data from parquet files from specified raw data path
}
