% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{scrape_inso_table}
\alias{scrape_inso_table}
\title{Scrape Insolvency Table}
\usage{
scrape_inso_table(url, date, item, search_param, url_result)
}
\arguments{
\item{url}{The URL for the initial search POST request.}

\item{date}{The date string being crawled (for logging).}

\item{item}{The 'gegenstand' item code.}

\item{search_param}{The search parameters for the initial POST request.}
}
\value{
A data frame containing the scraped data.
}
\description{
Scrapes the insolvency table for a given date and item, including fetching detailed text for each entry.
}
