% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{split_date}
\alias{split_date}
\title{Split Date}
\usage{
split_date(date)
}
\arguments{
\item{date}{A date object or a date string (e.g., "YYYY-MM-DD").}
}
\value{
A list of year, month, and day components as strings.
}
\description{
Splits a date object or string into year, month, and day components as strings.
}
