% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_data.R
\name{update_raw_data}
\alias{update_raw_data}
\title{Update Raw Insolvency Data}
\usage{
update_raw_data(data_path = "data-raw/inso")
}
\arguments{
\item{data_path}{The file path to the Parquet dataset directory.}
}
\description{
Orchestrates the entire data crawling process: loads existing data, determines dates to crawl,
performs scraping, cleans data, and saves updated results.
}
